--story=1003954 --user=lvzk 工单中心后端重构-对接工单中心查询-回显字段 https://www.tapd.cn/54247054/s/1130534

+++ src/main/java/codedriver/framework/process/workcenter/dto/WorkcenterVo.java
@@ -64,7 +64,7 @@ public class WorkcenterVo extends ConditionConfigVo implements Serializable {
 
     private List<WorkcenterTheadVo> theadVoList;
 
-    private List<Integer> processTaskIdList;
+    private List<Long> processTaskIdList;
 
     private String sqlFieldType;
 
@@ -272,11 +272,11 @@ public class WorkcenterVo extends ConditionConfigVo implements Serializable {
         this.sqlFieldType = sqlFieldType;
     }
 
-    public List<Integer> getProcessTaskIdList() {
+    public List<Long> getProcessTaskIdList() {
         return processTaskIdList;
     }
 
-    public void setProcessTaskIdList(List<Integer> processTaskIdList) {
+    public void setProcessTaskIdList(List<Long> processTaskIdList) {
         this.processTaskIdList = processTaskIdList;
     }
